# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.CreateTransactionRequest:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      category_id:
        type: integer
      currency:
        type: string
      date:
        type: string
      description:
        type: string
      transaction_type_id:
        type: integer
    type: object
  requests.CreateUserRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      userName:
        type: string
    type: object
...
  responses.CreateUserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      token:
        type: string
    type: object
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  responses.TransactionResponse:
    properties:
      account_id:
        type: integer
      amount:
        type: number
      category:
        type: string
      currency:
        type: string
      date:
        type: integer
      description:
        type: string
      id:
        type: integer
      transaction_type:
        type: string
    type: object
  responses.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Maybe Finance API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Maybe Finance API
  version: "1.0"
paths:
...
  /data/{account_id}/transactions/histogram:
    get:
      description: Get transaction histogram data
      operationId: get-transaction-histogram-data
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      summary: Get transaction histogram data
      tags:
      - data
  /data/{account_id}/transactions/summary:
    get:
      description: Get transactions summary data
      parameters:
      - description: Filter
        enum:
        - transaction_category
        - transaction_type
        - month
        - year
        - week
        in: query
        name: filter
        type: string
      - description: Interval
        enum:
        - month
        - year
        - week
        in: query
        name: interval
        type: string
      - description: Group by
        in: query
        name: group_by
        type: string
      - description: Sort by
        in: query
        name: sort_by
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items: {}
            type: array
      summary: Get transactions summary data
      tags:
      - data
...
  /transactions/{id}:
    get:
      description: Retrieve a transaction
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.TransactionResponse'
      summary: Get a transaction
      tags:
      - transactions
  /transactions/create:
    post:
      consumes:
      - application/json
      description: Create a transaction
      parameters:
      - description: Create Transaction Request
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/requests.CreateTransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a transaction
      tags:
      - transactions
...
  /users/{id}:
    get:
      description: Retrieve a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UserResponse'
      summary: Get a user
      tags:
      - users
  /users/create:
    post:
      consumes:
      - application/json
      description: Create a user
      parameters:
      - description: Create User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/requests.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.CreateUserResponse'
      summary: Create a user
      tags:
      - users
...
swagger: "2.0"